+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Unreleased in the current development version:
 - CLI script for automatic multiple experiment analysis submission (#1160, #1175)
 
 AQUA core complete list:
-- AQUA entry point for installation and catalog addition (#1131, #1134, #1146, #1168)
+- AQUA entry point for installation and catalog maintanance and fixes/grids handling (#1131, #1134, #1146, #1168, #1169)
 
 ## [v0.8.2]